+++ xenon-common/src/test/java/com/vmware/xenon/common/TestServiceHost.java
@@ -32,6 +32,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Random;
 import java.util.UUID;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.TimeUnit;
@@ -71,8 +72,10 @@ import com.vmware.xenon.services.common.MinimalFactoryTestService;
 import com.vmware.xenon.services.common.MinimalTestService;
 import com.vmware.xenon.services.common.NodeGroupService.NodeGroupState;
 import com.vmware.xenon.services.common.NodeState;
+import com.vmware.xenon.services.common.QueryTask.Query;
 import com.vmware.xenon.services.common.ServiceHostManagementService;
 import com.vmware.xenon.services.common.ServiceUriPaths;
+import com.vmware.xenon.services.common.UserService;
 
 public class TestServiceHost {
 
@@ -1347,6 +1350,29 @@ public class TestServiceHost {
         this.host.testWait();
     }
 
+    //override setProcessingStage() of ExampleService to randomly
+    // fail some pause operations
+    static class PauseExampleService extends ExampleService {
+
+        public static final String FACTORY_LINK = ServiceUriPaths.CORE + "/pause-examples";
+        public static final String STAT_NAME_ABORT_COUNT = "abortCount";
+
+        public static FactoryService createFactory() {
+            return FactoryService.create(PauseExampleService.class);
+        }
+
+        @Override
+        public void setProcessingStage(Service.ProcessingStage stage) {
+            if (stage == Service.ProcessingStage.PAUSED) {
+                if (new Random().nextBoolean()) {
+                    this.adjustStat(STAT_NAME_ABORT_COUNT, 1);
+                    throw new IllegalStateException("Cannot pause service.");
+                }
+            }
+            super.setProcessingStage(stage);
+        }
+    }
+
     @Test
     public void servicePauseDueToMemoryPressure() throws Throwable {
         setUp(true);
@@ -1378,6 +1404,31 @@ public class TestServiceHost {
         assertTrue(delayMicros == delayMicrosAfter);
         this.host.start();
 
+        this.host.setSystemAuthorizationContext();
+        TestContext ctx = this.host.testCreate(1);
+        String user = "foo@bar.com";
+        Query.Builder queryBuilder = Query.Builder.create()
+                .addFieldClause(ServiceDocument.FIELD_NAME_KIND, Utils.buildKind(ExampleServiceState.class));
+        AuthorizationSetupHelper.create()
+                .setHost(this.host)
+                .setUserEmail(user)
+                .setUserSelfLink(user)
+                .setUserPassword(user)
+                .setResourceQuery(queryBuilder.build())
+                .setCompletion((ex) -> {
+                    if (ex != null) {
+                        ctx.failIteration(ex);
+                    }
+                    ctx.completeIteration();
+                })
+                .start();
+        ctx.await();
+        this.host.startFactory(PauseExampleService.class,
+                PauseExampleService::createFactory);
+        URI factoryURI = UriUtils.buildFactoryUri(this.host, PauseExampleService.class);
+        this.host.waitForServiceAvailable(PauseExampleService.FACTORY_LINK);
+        this.host.resetSystemAuthorizationContext();
+
         AtomicLong selfLinkCounter = new AtomicLong();
         String prefix = "instance-";
         String name = UUID.randomUUID().toString();
@@ -1389,8 +1440,8 @@ public class TestServiceHost {
         };
 
         // Create a number of child services.
-        URI factoryURI = UriUtils.buildFactoryUri(this.host, ExampleService.class);
-        this.host.setSystemAuthorizationContext();
+        this.host.assumeIdentity(UriUtils.buildUriPath(UserService.FACTORY_LINK, user));
+
         Map<URI, ExampleServiceState> states = this.host.doFactoryChildServiceStart(null,
                 this.serviceCount,
                 ExampleServiceState.class, bodySetter, factoryURI);
@@ -1408,12 +1459,6 @@ public class TestServiceHost {
         if (this.testDurationSeconds > 0 || this.host.isStressTest()) {
             updateCount = 1;
         }
-        patchExampleServices(states, updateCount);
-
-        // Let's set the service memory limit back to normal and issue more updates to ensure
-        // that the services still continue to operate as expected.
-        this.host
-                .setServiceMemoryLimit(ServiceHost.ROOT_PATH, ServiceHost.DEFAULT_PCT_MEMORY_LIMIT);
         patchExampleServices(states, updateCount);
 
         this.host.testStart(states.size());
@@ -1436,12 +1481,17 @@ public class TestServiceHost {
                             });
             this.host.send(get);
         }
-
         this.host.testWait();
 
         if (this.testDurationSeconds == 0) {
             verifyPauseResumeStats(states);
         }
+
+        // Let's set the service memory limit back to normal and issue more updates to ensure
+        // that the services still continue to operate as expected.
+        this.host
+                .setServiceMemoryLimit(ServiceHost.ROOT_PATH, ServiceHost.DEFAULT_PCT_MEMORY_LIMIT);
+        patchExampleServices(states, updateCount);
 
         states.clear();
         // Long running test. Keep adding services, expecting pause to occur and free up memory so the
@@ -1479,7 +1529,7 @@ public class TestServiceHost {
             for (int i = 0; i < count; i += step) {
                 // now that we have created a bunch of services, and a lot of them are paused, ping one randomly
                 // to make sure it resumes
-                URI instanceUri = UriUtils.buildFactoryUri(this.host, ExampleService.class);
+                URI instanceUri = UriUtils.buildFactoryUri(this.host, PauseExampleService.class);
                 instanceUri = UriUtils.extendUri(instanceUri, prefix + (selfLinkCounter.get() - i));
                 Operation get = Operation.createGet(instanceUri).setCompletion((o, e) -> {
                     if (e == null) {
@@ -1506,60 +1556,30 @@ public class TestServiceHost {
         }
     }
 
+
     private void verifyPauseResumeStats(Map<URI, ExampleServiceState> states) throws Throwable {
         // Let's now query stats for each service. We will use these stats to verify that the
         // services did get paused and resumed.
         WaitHandler wh = () -> {
+            int totalServicePauseResumeOrAbort = 0;
             // Verify the stats for each service show that the service was paused and resumed
             for (ExampleServiceState st : states.values()) {
                 URI serviceUri = UriUtils.buildUri(this.host, st.documentSelfLink);
                 Map<String, ServiceStat> serviceStats = this.host.getServiceStats(serviceUri);
 
                 ServiceStat pauseStat = serviceStats.get(Service.STAT_NAME_PAUSE_COUNT);
-                if (pauseStat == null) {
-                    this.host.log("pauseStat was null for %s", st.documentSelfLink);
-                    return false;
-                }
-
                 ServiceStat resumeStat = serviceStats.get(Service.STAT_NAME_RESUME_COUNT);
-                if (resumeStat == null) {
-                    this.host.log("resumeStat was null for %s", st.documentSelfLink);
+                ServiceStat abortStat = serviceStats.get(PauseExampleService.STAT_NAME_ABORT_COUNT);
+                if (abortStat == null && (pauseStat == null || resumeStat == null)) {
                     return false;
                 }
+                totalServicePauseResumeOrAbort++;
             }
 
-            // Verify the stats for the management service show that the pause and resume service count
-            // is as expected
-            Map<String, ServiceStat> mgmtStats = this.host.getServiceStats(this.host
-                    .getManagementServiceUri());
-
-            ServiceStat mgmtPauseStat = mgmtStats.get(
-                    ServiceHostManagementService.STAT_NAME_SERVICE_PAUSE_COUNT);
-
-            ServiceStat mgmtResumeStat = mgmtStats.get(
-                    ServiceHostManagementService.STAT_NAME_SERVICE_PAUSE_COUNT);
-
-            double totalPauseResumeCount = mgmtPauseStat != null ? mgmtPauseStat.latestValue : 0L;
-            totalPauseResumeCount += mgmtResumeStat != null ? mgmtResumeStat.latestValue : 0L;
-
-            if (mgmtPauseStat == null) {
-                this.host.log("ManagementSvc pauseStat was not set as expected %s",
-                        mgmtPauseStat == null ? "null" : String.valueOf(mgmtPauseStat.latestValue));
-                return false;
-            }
-
-            if (mgmtResumeStat == null) {
+            if (totalServicePauseResumeOrAbort < states.size()) {
                 this.host.log(
-                        "ManagementSvc resumeStat was not set as expected %s",
-                        mgmtResumeStat == null ? "null" : String
-                                .valueOf(mgmtResumeStat.latestValue));
-                return false;
-            }
-
-            if (totalPauseResumeCount < states.size()) {
-                this.host.log(
-                        "ManagementSvc total pause + resume was less than service count %f (%d)",
-                        totalPauseResumeCount, states.size());
+                        "ManagementSvc total pause + resume or abort was less than service count %f (%d)",
+                        totalServicePauseResumeOrAbort, states.size());
                 return false;
             }